requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0
APScheduler>=3.10.0
